refactor: Extract CEL compiler to shared/pkg/cel/
Moves the CEL compiler implementation from services/reference-data/cel/
to shared/pkg/cel/ to enable reuse by Party service and gateway mapping
engine without circular dependencies.

- Created shared/pkg/cel/compiler.go with full compiler implementation
- Moved all test files (compiler, benchmark, determinism, regression, security) to shared/pkg/cel/
- Replaced services/reference-data/cel/compiler.go with backwards-compatible re-exports using type aliases and var re-exports
- All existing importers continue to work unchanged via re-exports
