+++ shared/pkg/cel/compiler_test.go
